Remove jobs that have been deleted in Sonar. Plus overall performance improvements for collectors.

+++ jenkins-build-collector/src/main/java/com/capitalone/dashboard/collector/HudsonCollectorTask.java
@@ -15,6 +15,7 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.stereotype.Component;
+import org.springframework.util.CollectionUtils;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -72,21 +73,27 @@ public class HudsonCollectorTask extends CollectorTask<HudsonCollector> {
     @Override
     public void collect(HudsonCollector collector) {
         long start = System.currentTimeMillis();
+        Set<ObjectId> udId = new HashSet<>();
+        udId.add(collector.getId());
+        List<HudsonJob> existingJobs = hudsonJobRepository.findByCollectorIdIn(udId);
+        List<HudsonJob> latestJobs = new ArrayList<>();
 
-        clean(collector);
+        clean(collector, existingJobs);
         for (String instanceUrl : collector.getBuildServers()) {
             logBanner(instanceUrl);
 
             Map<HudsonJob, Set<Build>> buildsByJob = hudsonClient
                     .getInstanceJobs(instanceUrl);
             log("Fetched jobs", start);
-
-            addNewJobs(buildsByJob.keySet(), collector);
+            latestJobs.addAll(buildsByJob.keySet());
+            addNewJobs(buildsByJob.keySet(), existingJobs, collector);
 
             addNewBuilds(enabledJobs(collector, instanceUrl), buildsByJob);
 
             log("Finished", start);
         }
+        // First delete jobs that will be no longer collected because servers have moved etc.
+        deleteUnwantedJobs(latestJobs, existingJobs, collector);
 
     }
 
@@ -94,13 +101,13 @@ public class HudsonCollectorTask extends CollectorTask<HudsonCollector> {
      * Clean up unused hudson/jenkins collector items
      *
      * @param collector the {@link HudsonCollector}
+     * @param existingJobs
      */
 
     @SuppressWarnings("PMD.UnnecessaryFullyQualifiedName")
-    private void clean(HudsonCollector collector) {
+    private void clean(HudsonCollector collector, List<HudsonJob> existingJobs) {
 
-        // First delete jobs that will be no longer collected because servers have moved etc.
-        deleteUnwantedJobs(collector);
+
         Set<ObjectId> uniqueIDs = new HashSet<>();
         for (com.capitalone.dashboard.model.Component comp : dbComponentRepository
                 .findAll()) {
@@ -117,32 +124,36 @@ public class HudsonCollectorTask extends CollectorTask<HudsonCollector> {
 
             }
         }
-        List<HudsonJob> jobList = new ArrayList<>();
+        List<HudsonJob> stateChangeJobList = new ArrayList<>();
         Set<ObjectId> udId = new HashSet<>();
         udId.add(collector.getId());
-        for (HudsonJob job : hudsonJobRepository.findByCollectorIdIn(udId)) {
-            if (job != null) {
+        for (HudsonJob job : existingJobs) {
+            if ((job.isEnabled() && !uniqueIDs.contains(job.getId())) ||  // if it was enabled but not on a dashboard
+                    (!job.isEnabled() && uniqueIDs.contains(job.getId()))) { // OR it was disabled and now on a dashboard
                 job.setEnabled(uniqueIDs.contains(job.getId()));
-                jobList.add(job);
+                stateChangeJobList.add(job);
             }
         }
-        hudsonJobRepository.save(jobList);
+        if (!CollectionUtils.isEmpty(stateChangeJobList)) {
+            hudsonJobRepository.save(stateChangeJobList);
+        }
     }
 
-    private void deleteUnwantedJobs(HudsonCollector collector) {
+    private void deleteUnwantedJobs(List<HudsonJob> latestJobs, List<HudsonJob> existingJobs, HudsonCollector collector) {
 
         List<HudsonJob> deleteJobList = new ArrayList<>();
         Set<ObjectId> udId = new HashSet<>();
         udId.add(collector.getId());
-        for (HudsonJob job : hudsonJobRepository.findByCollectorIdIn(udId)) {
+        for (HudsonJob job : existingJobs) {
             if (!collector.getBuildServers().contains(job.getInstanceUrl()) ||
-                    (!job.getCollectorId().equals(collector.getId()))) {
+                    (!job.getCollectorId().equals(collector.getId())) ||
+                    (!latestJobs.contains(job))) {
                 deleteJobList.add(job);
             }
         }
-
-        hudsonJobRepository.delete(deleteJobList);
-
+        if (!CollectionUtils.isEmpty(deleteJobList )) {
+            hudsonJobRepository.delete(deleteJobList);
+        }
     }
 
     /**
@@ -181,25 +192,28 @@ public class HudsonCollectorTask extends CollectorTask<HudsonCollector> {
 
     /**
      * Adds new {@link HudsonJob}s to the database as disabled jobs.
-     *
-     * @param jobs      list of {@link HudsonJob}s
+     *  @param jobs      list of {@link HudsonJob}s
+     * @param existingJobs
      * @param collector the {@link HudsonCollector}
      */
-    private void addNewJobs(Set<HudsonJob> jobs, HudsonCollector collector) {
+    private void addNewJobs(Set<HudsonJob> jobs, List<HudsonJob> existingJobs, HudsonCollector collector) {
         long start = System.currentTimeMillis();
         int count = 0;
 
+        List<HudsonJob> newJobs = new ArrayList<>();
         for (HudsonJob job : jobs) {
-
-            if (isNewJob(collector, job)) {
+            if (!existingJobs.contains(job)) {
                 job.setCollectorId(collector.getId());
                 job.setEnabled(false); // Do not enable for collection. Will be
                 // enabled when added to dashboard
                 job.setDescription(job.getJobName());
-                hudsonJobRepository.save(job);
+                newJobs.add(job);
                 count++;
             }
-
+        }
+        //save all in one shot
+        if (!CollectionUtils.isEmpty(newJobs)) {
+            hudsonJobRepository.save(newJobs);
         }
         log("New jobs", start, count);
     }
